jsop.py: use 0xff to seperate UTF8 strings
So far, we used 0x00 as the seperator between the strings, that were
encoded in UTF8. For this reason, it was impossible to use a string
that contained U+0000 as a map key, even though it's technically a
legal unicode character, and this can be done in JSON.

Note that this is a non-backward-compatible change in the file format.