Garbage Collection Bug Fix
Up until now, when a map or a list were deleted, their values were also
deleted, but the pointer to the first key (always null), the pointer to the
last key (always null), and the size (always 0) would not be deleted.

Also, this commit seperates the JObject.clear() method, which is part of
the API, from the new JObject._destroy() method, which is internal.

In the same spirit, the method JObject._init() is renamed to
JObject._init(), which is more appropriate, since it is not intended for
external use.